Support underline colors and styles in non-gui mode
Many terminal emulators now support informal escape sequences
to allow setting color and style of underlines.  Underline color is
typically controlled by a sequence intended to mirror the CSI 38/48
codes for foreground/background colors:

    CSI 58 : 2 :: R : G : B m   -> set color to specified true color RGB
    CSI 58 : 5 : I m            -> set color to palette index I (0-255)
    CSI 59 m                    -> restore underline color to default

Underline style is controlled via a set of extensions to CSI 4:

    CSI 4 : 0 m   -> No underline
    CSI 4 : 1 m   -> Single underline
    CSI 4 : 2 m   -> Double underline
    CSI 4 : 3 m   -> Curly underline
    CSI 4 : 4 m   -> Dotted underline
    CSI 4 : 5 m   -> Dashed underline

The informal terminfo identifiers for these features are:

    Smulx  -> Underline style sequence
    Setulc -> Underline color set sequence
    ol     -> Underline color reset sequence

This patch allows emacs to recognize these features of a terminal
emulator and use them to support setting 'wave' underline style
and colors when running in a tty.

(Bug#45230)
(Bug#45231)